Organizes algo functions into separate files
