+++ src/algorithms/_engine.ts
@@ -3,7 +3,7 @@ import { EventEmitter } from 'events'
 import EventBus from '../__core/event-bus'
 import { Logger } from '../__core/logger'
 import { TransactionPacket, Tick as Tick, RequestParams, ChartData, QueueType, ExtTick } from '../types/types'
-import { algo1 } from './algo-test-1'
+import { algoOutput } from './_output'
 
 export class Queue {
   private bus: EventEmitter
@@ -35,7 +35,7 @@ export class Queue {
           queueData.inputType,
           ...queueData.queue.slice(0, 2),
         )
-        const algoResult = algo1(requestParams, queueData)
+        const algoResult = algoOutput(requestParams, queueData)
         this.bus.emit('analysisResults:data', algoResult, chartData)
       },
     )
